Moved v2 team components to features
Adds deprecation notices
